// WebAPIAngular.Angular/src/app/model/femaSubIndex.ts
export class FemaSubIndex {
    SubIndexId: number;
    IndexId: number;
    SubIndexNo: string;
    SubIndexName: string;
    SubIndexContent: string;
    SortId: number;
    IsDeleted: boolean;
    IsActive: boolean;
    CreatedDate: string;
    ModifiedDate: string;
}

export class GetFemaSubIndexRequest {
    SubIndexId: number;
    IndexId: number;
    RegulationId: number;
    SearchText: string;
    IsActive: boolean;
    PageNumber: number;
    PageSize: number;
    OrderBy: string;
    OrderByDirection: string;
}